requirements.txt
fastapi
uvicorn[standard]
# openai
# azure-search-documents
azure-core
python-multipart
python-dotenv
python-docx
PyPDF2
openai>=1.3.8,<2
azure-search-documents>=11.4.0
pydantic

